+++ tests/test_account_logout.py
@@ -3,9 +3,8 @@ from curl import main_site, main_profile, main_login
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from locators import Locators
-from data import AuthGenerator
+from data import email_register, password_register
 
-auth_generator = AuthGenerator()
 
 
 @pytest.mark.usefixtures("driver")
@@ -17,8 +16,8 @@ class TestLogoutTest:
         )
         driver.find_element(*Locators.PROFILE_LINK).click()
 
-        email = auth_generator.email_register
-        password = auth_generator.password_register
+        email = email_register
+        password = password_register
 
         driver.find_element(*Locators.NAME).send_keys(email)
         driver.find_element(*Locators.PASSWORD_INPUT).send_keys(password)
