+++ tests/test_registration.py
@@ -3,7 +3,7 @@ from curl import main_site,main_login
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from locators import Locators
-from data import AuthGenerator
+from helpers import AuthGenerator
 
 
 @pytest.mark.usefixtures("driver")
